+++ src/qt/walletmodel.h
@@ -63,10 +63,10 @@ public:
     AddressTableModel *getAddressTableModel();
     TransactionTableModel *getTransactionTableModel();
 
-    qint64 getBalance(const CCoinControl *coinControl=NULL) const;
-    qint64 getStake() const;
-    qint64 getUnconfirmedBalance() const;
-    qint64 getImmatureBalance() const;
+    __int128 getBalance(const CCoinControl *coinControl=NULL) const;
+    __int128 getStake() const;
+    __int128 getUnconfirmedBalance() const;
+    __int128 getImmatureBalance() const;
     EncryptionStatus getEncryptionStatus() const;
 
     // Check address for validity
@@ -137,10 +137,10 @@ private:
     TransactionTableModel *transactionTableModel;
 
     // Cache some values to be able to detect changes
-    qint64 cachedBalance;
-    qint64 cachedStake;
-    qint64 cachedUnconfirmedBalance;
-    qint64 cachedImmatureBalance;
+    __int128 cachedBalance;
+    __int128 cachedStake;
+    __int128 cachedUnconfirmedBalance;
+    __int128 cachedImmatureBalance;
     EncryptionStatus cachedEncryptionStatus;
     int cachedNumBlocks;
 
@@ -163,7 +163,7 @@ public slots:
 
 signals:
     // Signal that balance in wallet changed
-    void balanceChanged(qint64 balance, qint64 stake, qint64 unconfirmedBalance, qint64 immatureBalance);
+    void balanceChanged(__int128 balance, __int128 stake, __int128 unconfirmedBalance, __int128 immatureBalance);
 
     // Encryption status of wallet changed
     void encryptionStatusChanged(int status);
